Extract QueryConfig from BaseJdbcConfig
Conditional module requires to populate BaseJdbcConfig. Because of the
fact some jdbc connectors overrides defaults of BaseJdbcConfig, these
defaults are not visible during mentioned population and so that could
lead to validation error of the config.

Consider connectors that are not using connection-url config property.
